+++ internal/api/design/ingest.go
@@ -17,7 +17,7 @@ var _ = Service("ingest", func() {
 		Response("forbidden", StatusForbidden)
 	})
 	Method("monitor_request", func() {
-		Description("Request access to the /monitor WebSocket")
+		Description("Request access to the /monitor SSE event stream")
 		// Do not require any scope, user claims will be stored internally
 		// and checked in the monitor endpoint after validating the cookie.
 		BearerAuthScopes()
@@ -32,7 +32,7 @@ var _ = Service("ingest", func() {
 			POST("/monitor")
 			Response("internal_error", StatusInternalServerError)
 			Response(StatusOK, func() {
-				Cookie("ticket:enduro-ingest-ws-ticket")
+				Cookie("ticket:enduro-ingest-sse-ticket")
 				CookieMaxAge(5)
 				CookieSecure()
 				CookieHTTPOnly()
@@ -40,7 +40,7 @@ var _ = Service("ingest", func() {
 		})
 	})
 	Method("monitor", func() {
-		Description("Obtain access to the /monitor WebSocket")
+		Description("Obtain access to the /monitor SSE event stream")
 		// Disable BearerAuth security (it validates the previous method's cookie).
 		NoSecurity()
 		Payload(func() {
@@ -50,9 +50,9 @@ var _ = Service("ingest", func() {
 		Error("internal_error")
 		HTTP(func() {
 			GET("/monitor")
+			ServerSentEvents()
 			Response("internal_error", StatusInternalServerError)
-			Response(StatusOK)
-			Cookie("ticket:enduro-ingest-ws-ticket")
+			Cookie("ticket:enduro-ingest-sse-ticket")
 		})
 	})
 	Method("list_sips", func() {
@@ -669,7 +669,7 @@ var SIPTask = ResultType("application/vnd.enduro.ingest.sip.task", func() {
 		Attribute("note", String)
 		TypedAttributeUUID("workflow_uuid", "Identifier of related workflow")
 	})
-	Required("uuid", "name", "status", "started_at", "workflow_uuid")
+	Required("uuid", "name", "status", "workflow_uuid")
 })
 
 var SIPDecision = ResultType("application/vnd.enduro.ingest.sip.decision", func() {